[VISITOR] Base implementation (non-reflection) - Case 1

+++ Assets/Design Patterns/Visitor/Case Study 1 - Shape Calculators/Implementation C# - Extension Method/Scripts [Visitor - Calculator]/Calculator.cs
@@ -5,18 +5,18 @@ namespace VisitorPattern.Case1.CSharp {
   /// * [The 'Abstract Visitor' class]
   /// </summary>
   public abstract class Calculator {
-    public float Calculate(ProceduralShape shape) {
+    public double Calculate(ProceduralShape shape) {
       var shapeType = shape.GetType();
       var calculatorType = this.GetType();
       var calculatingMethod = calculatorType.GetNonPublicMethod(nameof(Calculate), paramType: shapeType);
 
-      return (float)calculatingMethod?.Invoke(this, new[] { shape });
+      return (double)calculatingMethod?.Invoke(this, new[] { shape });
 
       throw new NotImplementedException();
     }
 
-    protected abstract float Calculate(ProceduralCube shape);
-    protected abstract float Calculate(ProceduralSphere shape);
+    protected abstract double Calculate(ProceduralCube shape);
+    protected abstract double Calculate(ProceduralSphere shape);
 
     // ! If visitor's operation is not mandatory for all elements, declare virtual methods instead:
     // private SystemException GetNotImplementedException(ICalculatable element) {
